make package eventhandler a subpackage of events. Move existing eventHandler into this package
