+++ contribs/freight/src/main/java/org/matsim/contrib/freight/events/eventhandler/LSPLinkLeaveEventHandler.java
@@ -1,4 +1,4 @@
-package org.matsim.contrib.freight.eventhandler;
+package org.matsim.contrib.freight.events.eventhandler;
 
 import org.matsim.contrib.freight.events.LSPFreightLinkLeaveEvent;
 import org.matsim.core.events.handler.EventHandler;
