+++ contribs/freight/src/main/java/org/matsim/contrib/freight/events/eventhandler/LSPServiceStartEventHandler.java
@@ -1,4 +1,4 @@
-package org.matsim.contrib.freight.eventhandler;
+package org.matsim.contrib.freight.events.eventhandler;
 
 import org.matsim.contrib.freight.events.LSPServiceStartEvent;
 import org.matsim.core.events.handler.EventHandler;
