+++ contribs/freight/src/main/java/org/matsim/contrib/freight/events/eventhandler/ShipmentPickedUpEventHandler.java
@@ -1,5 +1,6 @@
-package org.matsim.contrib.freight.events;
+package org.matsim.contrib.freight.events.eventhandler;
 
+import org.matsim.contrib.freight.events.ShipmentPickedUpEvent;
 import org.matsim.core.events.handler.EventHandler;
 
 /**
